feat(neovim): Make window borders stand out
Ref:
    https://github.com/folke/tokyonight.nvim/issues/34

With the current config, the window borders are not visible at all,
which is very bad. I read the documentation for LazyVim, Neovim, Vim
but could not solve the problem. In the end, I solved the problem by
customizing the tokyonight theme. See above issue link for detail.

I had a very had time solving this problem :(
(I am not familiar with Neovim and Vim)